feat: add support for eager imports via ImportGlobOptions argument
According to Vite types, globEager is deprecated in favour of supplying an
options argument to glob (from importMeta.d.ts 👇🏽)

```
/**
 * @deprecated Use `import.meta.glob('*', { eager: true })` instead
 */
globEager: import('./importGlob').ImportGlobEagerFunction
```

This commit adds tests cases to cover the desired behaviour, and the
appropriate implementation. The target code was obtained from the Vite
docs at [1]. This only supports the eager option, others are not covered
here.

[1] https://vitejs.dev/guide/features.html#glob-import:~:text=%7B%20eager%3A%20true%20%7D)-,The%20above%20will%20be%20transformed%20into%20the%20following%3A,-js